More changes for the show post page
