+++ locale/de/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Benutzerdefiniert",
   "label.description": "Beschreibung",
   "label.discussion": "Diskussion",
+  "label.edittags": "",
   "label.email": "E-Mail",
   "label.follow": "",
   "label.gravatar": "Gravatar",
