+++ locale/el/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Προσαρμοσμένη",
   "label.description": "Περιγραφή",
   "label.discussion": "Συζήτηση",
+  "label.edittags": "",
   "label.email": "Email",
   "label.follow": "",
   "label.gravatar": "Gravatar",
