+++ locale/en/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Custom",
   "label.description": "Description",
   "label.discussion": "Discussion",
+  "label.edittags": "Edit tags",
   "label.email": "Email",
   "label.follow": "Follow",
   "label.gravatar": "Gravatar",
