+++ locale/es-ES/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Personalizado",
   "label.description": "Descripción",
   "label.discussion": "Discusión",
+  "label.edittags": "",
   "label.email": "Correo electrónico",
   "label.follow": "",
   "label.gravatar": "Gravatar",
