+++ locale/fr/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Personnalisé",
   "label.description": "Description",
   "label.discussion": "Discussion",
+  "label.edittags": "",
   "label.email": "E-mail",
   "label.follow": "",
   "label.gravatar": "Gravatar",
