+++ locale/it/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Personalizzato",
   "label.description": "Descrizione",
   "label.discussion": "Discussione",
+  "label.edittags": "",
   "label.email": "Email",
   "label.follow": "",
   "label.gravatar": "Gravatar",
