+++ locale/nl/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Aangepast",
   "label.description": "Omschrijving",
   "label.discussion": "Discussie",
+  "label.edittags": "",
   "label.email": "E-mailadres",
   "label.follow": "",
   "label.gravatar": "Gravatar",
