+++ locale/pl/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Niestandardowy",
   "label.description": "Opis",
   "label.discussion": "Dyskusja",
+  "label.edittags": "",
   "label.email": "Email",
   "label.follow": "",
   "label.gravatar": "Gravatar",
