+++ locale/pt-BR/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Personalizado",
   "label.description": "Descrição",
   "label.discussion": "Discussão",
+  "label.edittags": "",
   "label.email": "E-mail",
   "label.follow": "",
   "label.gravatar": "Gravatar",
