+++ locale/sk/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Vlastné",
   "label.description": "Popis",
   "label.discussion": "Diskusia",
+  "label.edittags": "",
   "label.email": "Email",
   "label.follow": "",
   "label.gravatar": "Gravatar",
