+++ locale/sv-SE/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Anpassad",
   "label.description": "Beskrivning",
   "label.discussion": "Diskussion",
+  "label.edittags": "",
   "label.email": "E-post",
   "label.follow": "",
   "label.gravatar": "Gravatar",
