+++ locale/tr/client.json
@@ -54,6 +54,7 @@
   "label.custom": "Özel",
   "label.description": "Açıklama",
   "label.discussion": "Tartışma",
+  "label.edittags": "",
   "label.email": "E-Posta",
   "label.follow": "",
   "label.gravatar": "Gravatar",
