+++ locale/zh-CN/client.json
@@ -54,6 +54,7 @@
   "label.custom": "自定义",
   "label.description": "说明",
   "label.discussion": "讨论",
+  "label.edittags": "",
   "label.email": "邮箱",
   "label.follow": "",
   "label.gravatar": "Gravatar",
