+++ public/components/ShowTag.scss
@@ -4,12 +4,13 @@
   display: inline-flex;
   justify-content: center;
   align-items: center;
-  font-size: 11px;
+  font-size: 13px;
   color: get("colors.gray.900");
-  font-weight: 500;
-  border-radius: 10px;
-  line-height: 18px;
-  padding: 0 8px;
+  background-color: get("colors.gray.200");
+  font-weight: 600;
+  border-radius: 5px;
+  // line-height: 18px;
+  padding: 6px 10px;
   border: 0;
 
   &[href] {
@@ -19,6 +20,14 @@
     &:hover {
       opacity: 1;
     }
+  }
+
+  span {
+    margin-right: 7px;
+    width: 12px;
+    height: 12px;
+    border-radius: 50%;
+    display: inline-block;
   }
 
   svg {
