+++ public/pages/ShowPost/components/FollowButton.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from "react"
-import { Post } from "@fider/models"
 import { Button, Icon } from "@fider/components"
 import { actions } from "@fider/services"
 import { useFider } from "@fider/hooks"
@@ -7,13 +6,14 @@ import IconVolumeOn from "@fider/assets/images/heroicons-volume-on.svg"
 import IconVolumeOff from "@fider/assets/images/heroicons-volume-off.svg"
 import { VStack } from "@fider/components/layout"
 import { Trans } from "@lingui/macro"
+import { Post } from "@fider/models"
 
 export interface NotificationsPanelProps {
   post: Post
   subscribed: boolean
 }
 
-export const NotificationsPanel = (props: NotificationsPanelProps) => {
+export const FollowButton = (props: NotificationsPanelProps) => {
   const fider = useFider()
   const [subscribed, setSubscribed] = useState(props.subscribed)
 
@@ -31,38 +31,20 @@ export const NotificationsPanel = (props: NotificationsPanelProps) => {
   }
 
   const button = subscribed ? (
-    <Button className="w-full" onClick={subscribeOrUnsubscribe} disabled={fider.isReadOnly}>
+    <Button variant="quaternary" className="w-full text-primary-base" onClick={subscribeOrUnsubscribe} disabled={fider.isReadOnly}>
       <Icon sprite={IconVolumeOff} />{" "}
       <span>
-        <Trans id="label.unsubscribe">Unsubscribe</Trans>
+        <Trans id="label.unfollow">Unfollow</Trans>
       </span>
     </Button>
   ) : (
-    <Button className="w-full" onClick={subscribeOrUnsubscribe} disabled={fider.isReadOnly}>
+    <Button variant="quaternary" className="w-full text-primary-base" onClick={subscribeOrUnsubscribe} disabled={fider.isReadOnly}>
       <Icon sprite={IconVolumeOn} />
       <span>
-        <Trans id="label.subscribe">Subscribe</Trans>
+        <Trans id="label.follow">Follow</Trans>
       </span>
     </Button>
   )
 
-  const text = subscribed ? (
-    <span className="text-muted">
-      <Trans id="showpost.notificationspanel.message.subscribed">You’re receiving notifications about activity on this post.</Trans>
-    </span>
-  ) : (
-    <span className="text-muted">
-      <Trans id="showpost.notificationspanel.message.unsubscribed">You&apos;ll not receive any notification about this post.</Trans>
-    </span>
-  )
-
-  return (
-    <VStack>
-      <span className="text-category">
-        <Trans id="label.notifications">Notifications</Trans>
-      </span>
-      {button}
-      {text}
-    </VStack>
-  )
+  return <VStack>{button}</VStack>
 }
